Add files via upload
Matplotlib
Matplotlib is a module that contains classes and functions for creating MATLAB style graphics and plots.

The primary submodule we will use is pyplot, which is commonly aliased as plt.

import matplotlib.pyplot as plt

Figure and Axis Objects
A figure is a base container that defines the canvas on which the plot appears. The axes object is the set of axes (usually x and y) that the data are plotted on.

Simple Plots¶
The quickest way to generate a simple, 1-D plot is using the pyplot.plot() function. pyplot.plot() automatically creates both the figure and axis objects and plots the data onto them.

Further reading:-
https://matplotlib.org/gallery.html#lines_bars_and_markers